Move builtin and fb_type IDs out of opcode.rs
opcode.rs held three separate things: the opcode encoding, the built-in
function IDs addressed through the BUILTIN operand, and the function block
type IDs addressed through FB_CALL. At 1431 lines it was also well over the
1000-line module limit.

Move the two ID modules to their own files verbatim and re-export them at
their existing paths, so opcode::builtin::EXPT_I32 and opcode::fb_type::TON
still resolve and no call site changes. No ID value changes; the only edits
to the moved text are the dedent and the reflow rustfmt applies to it.
